+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `MessageDescriptor::get_extension_by_full_name()`.
+
 ## [0.8.1] - 2022-05-29
 
 ### Added